improved 'improved' collisions
Random spacelaunch effect fixed.  vibration still an issue.  Walking
into walls now teleports you to the top of the wall, which is an
improvement, but not what I had in mind.  Need to write a CollisionMap
class with a Vector2[,] of points.  (Could be just a Vector[], but would
need a map. Which is doable.)